fixing forms errors

+++ resources/js/Components/MapsComponents/Activitie/TransmissionLine/MarkerManager.jsx
@@ -7,11 +7,14 @@ import TowerImportModal from "../../../TowersComponents/TowerImportModal";
 import "react-toastify/dist/ReactToastify.css";
 // import axios from "axios";
 import axios from "../../../../Components/axiosInstance";
+import getFirstErrorMessage from "../../../processLaravelErrors";
+import Swal from "sweetalert2";
 
 const MarkerManager = ({ show, onHide }) => {
     const [editedMarker, setEditedMarker] = useState(null);
     const [markers, setMarkers] = useState([]);
     const [configModalShow, setConfigModalShow] = useState(false);
+    const [errors, setErrors] = useState({});
 
     useEffect(() => {
         // Fetch markers when the component mounts
@@ -48,16 +51,21 @@ const MarkerManager = ({ show, onHide }) => {
 
             const response = await axios.post("/api/markers", formData);
 
-            if (response.status === 200) {
+            if (response.status === 200 || response.status === 201) {
                 const data = response.data;
                 updateMarkersList(data);
                 toast.success("Marcador salvo com sucesso");
+                setErrors({});
+                setConfigModalShow(false);
             } else {
-                handleSuccessMessage();
-                throw new Error("Erro ao salvar o marcador");
+                toast.error(response.statusText);
+                throw new Error("Erro ao salvar");
             }
         } catch (error) {
-            toast.error(error.message);
+            const message = getFirstErrorMessage(error.response.data);
+            // console.log(message)
+            setErrors(error.response.data);
+            // toast.error(message);
         }
     };
 
@@ -78,12 +86,17 @@ const MarkerManager = ({ show, onHide }) => {
                 const data = response.data;
                 updateMarkersList(data);
                 toast.success("Marcador atualizado com sucesso!");
+                setErrors({});
+                setConfigModalShow(false);
             } else {
                 toast.error(response.statusText);
-                throw new Error("Erro ao atualizar o marcador");
+                throw new Error("Erro ao atualizar");
             }
         } catch (error) {
-            toast.error(error.message);
+            const message = getFirstErrorMessage(error.response.data);
+            // console.log(message)
+            setErrors(error.response.data);
+            // toast.error(message);
         }
     };
 
@@ -108,19 +121,31 @@ const MarkerManager = ({ show, onHide }) => {
 
     const handleDeleteMarker = async (id) => {
         try {
-            const response = await axios.delete(`/api/markers/${id}`);
+            Swal.fire({
+                title: "Tem certeza?",
+                text: "Você não poderá reverter isso!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Sim, exclua!",
+            }).then(async (result) => {
+                if (result.isConfirmed) {
+                    const response = await axios.delete(`/api/markers/${id}`);
 
-            if (response.status === 200) {
-                setMarkers((prevMarkers) =>
-                    prevMarkers.filter((m) => m.id !== id),
-                );
-                toast.success("Deletado com sucesso!");
-                handleCloseModal();
-            } else {
-                throw new Error("Failed to delete marker");
-            }
+                    if (response.status === 200 || response.status === 204) {
+                        setMarkers((prevMarkers) =>
+                            prevMarkers.filter((m) => m.id !== id),
+                        );
+                        toast.success("Deletado com sucesso!");
+                        handleCloseModal();
+                    } else {
+                        throw new Error("Failed to delete marker");
+                    }
+                }
+            });
         } catch (error) {
-            toast.error("Error deleting marker: " + error.message);
+            toast.error(error.message);
         }
     };
 
@@ -158,6 +183,7 @@ const MarkerManager = ({ show, onHide }) => {
                         onSave={handleSaveMarker}
                         onUpdate={handleUpdateMarker}
                         editedMarker={editedMarker}
+                        errors={errors}
                     />
                 </Modal.Body>
                 <ToastContainer />
