+++ resources/js/Components/MapsComponents/Impediments/MarkerManagerImpediment.jsx
@@ -6,11 +6,14 @@ import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 // import axios from "axios";
 import axios from "../../../Components/axiosInstance";
+import getFirstErrorMessage from "../../processLaravelErrors";
+import Swal from "sweetalert2";
 
 const MarkerManagerImpediment = ({ show, onHide }) => {
     const [editedMarker, setEditedMarker] = useState(null);
     const [markers, setMarkers] = useState([]);
     const [configModalShow, setConfigModalShow] = useState(false);
+    const [errors, setErrors] = useState({});
 
     useEffect(() => {
         // Fetch markers when the component mounts
@@ -61,9 +64,13 @@ const MarkerManagerImpediment = ({ show, onHide }) => {
 
             updateMarkersList(response.data);
             toast.success("Marcador salvo com sucesso");
+            setErrors({});
+            setConfigModalShow(false);
         } catch (error) {
-            console.error("Erro ao salvar o marcador:", error.message);
-            toast.error("Erro ao salvar o marcador: " + error.message);
+            const message = getFirstErrorMessage(error.response.data);
+            // console.log(message)
+            setErrors(error.response.data);
+            // toast.error(message);
         }
     };
 
@@ -95,12 +102,17 @@ const MarkerManagerImpediment = ({ show, onHide }) => {
                 const data = response.data;
                 updateMarkersList(data);
                 toast.success("Marcador atualizado com sucesso!");
+                setErrors({});
+                setConfigModalShow(false);
             } else {
                 toast.error(response.statusText);
                 throw new Error("Erro ao atualizar o marcador");
             }
         } catch (error) {
-            toast.error(error.message);
+            const message = getFirstErrorMessage(error.response.data);
+            // console.log(message)
+            setErrors(error.response.data);
+            // toast.error(message);
         }
     };
 
@@ -125,21 +137,39 @@ const MarkerManagerImpediment = ({ show, onHide }) => {
 
     const handleDeleteMarker = async (id) => {
         try {
-            const response = await axios.delete(
-                `/api/markersimpediments/${id}`,
-            );
+            Swal.fire({
+                title: "Tem certeza?",
+                text: "Você não poderá reverter isso!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Sim, exclua!",
+            }).then(async (result) => {
+                if (result.isConfirmed) {
+                    const response = await axios.delete(
+                        `/api/markersimpediments/${id}`,
+                    );
 
-            if (response.status === 200) {
-                setMarkers((prevMarkers) =>
-                    prevMarkers.filter((m) => m.id !== id),
-                );
-                toast.success("Deletado com sucesso!");
-                handleCloseModal(); // Close modal after successful deletion
-            } else {
-                throw new Error("Failed to delete marker");
-            }
+                    if (response.status === 200 || response.status === 204) {
+                        setMarkers((prevMarkers) =>
+                            prevMarkers.filter((m) => m.id !== id),
+                        );
+                        toast.success("Excluído com sucesso!");
+
+                        Swal.fire({
+                            title: "Excluído!",
+                            text: "Seu registro foi excluído.",
+                            icon: "success",
+                        });
+                        // handleCloseModal();
+                    } else {
+                        throw new Error("Falha ao excluir");
+                    }
+                }
+            });
         } catch (error) {
-            toast.error("Error deleting marker: " + error.message);
+            toast.error(error.message);
         }
     };
 
@@ -179,6 +209,7 @@ const MarkerManagerImpediment = ({ show, onHide }) => {
                         onSave={handleSaveMarker}
                         onUpdate={handleUpdateMarker}
                         editedMarker={editedMarker}
+                        errors={errors}
                     />
                 </Modal.Body>
                 <ToastContainer />
